allure_lesson/requirements.txt
allure-pytest==2.14.1
pytest==8.3.3
pytest-xdist==3.6.1
